Completed concurrent SharedTable concurrent operations.
This commit contains a cache for the client which is updated on
puts and gets and concurrency controll to enable multiple clients
to access the cache. Furhtermore this commit also ensures that the
region server are not in an inconsistant state with rows just present
in a single regionserver. This problem could have arris if through
the execution of a protocol a value was being inserted and it was
only located on a single region. In this current implementation only
a single put operation can be executed at any given time and no
read operation can be performed. This concurrency control is
ensured with a readwrite lock.