Made database schema global and Optimized SharedTable.
SharedTable and SharedAdmin has until now replicated every requested to three clusters, making all of them identical. This was necessary has our goal is not data replication or availability but security, which caused unnecessary network requests and storage overheads. To solve this problem, the database schema has been moved to static variable which can be accessed by both classes to distinguish protected table from normal tables. The latter are stored only in a single cluster, the first one, while the latter are created on all of the clusters. Protected tables store all values with the assigned encrypted or just replicate the values if not encryption is defined. With this optimization, both network usage and storage are optimized to use only the necessary resources.